Refactor state management to support message propagation
This is the first step to more robust and easier support for more
complicated state changes. In Elm state can be controlled with events or
messages that can be returned by other events from update functions.
These returned events are usually called commands, which can be
functions or other more complex logic. In this refactoring I focused on
returning a single message or event first.

The nature of this refactoring resulted in a large commit that touches
many files. All the update logic not related directly to the main app
were moved under respective modules.

State management is now more aligned with the ratatui [Elm architecture
example](https://ratatui.rs/concepts/application-patterns/the-elm-architecture/).